# docs/swagger.yaml
basePath: /api
definitions:
  github_com_fisker_zjump-backend_internal_model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_fisker_zjump-backend_internal_model.SystemUser:
    properties:
      authType:
        description: 认证信息 (必须明确选择 password 或 key，不支持 auto)
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        description: 显示名称
        type: string
      passphrase:
        type: string
      password:
        type: string
      priority:
        description: 其他设置
        type: integer
      privateKey:
        type: string
      protocol:
        description: 协议和设置
        type: string
      status:
        description: active, inactive
        type: string
      updatedAt:
        type: string
      username:
        description: OS用户名
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zjump.com
    name: API Support
  description: ZJump 堡垒机系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ZJump API
  version: "2.0"
paths:
...
  /api/system-users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取系统用户列表
      tags:
      - system-users
    post:
      consumes:
      - application/json
      parameters:
      - description: System User
        in: body
        name: systemUser
        required: true
        schema:
          $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.SystemUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 创建系统用户
      tags:
      - system-users
  /api/system-users/{id}:
    delete:
      parameters:
      - description: System User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 删除系统用户
      tags:
      - system-users
    get:
      parameters:
      - description: System User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取单个系统用户
      tags:
      - system-users
    put:
      consumes:
      - application/json
      parameters:
      - description: System User ID
        in: path
        name: id
        required: true
        type: string
      - description: System User
        in: body
        name: systemUser
        required: true
        schema:
          $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.SystemUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 更新系统用户
      tags:
      - system-users
  /api/system-users/available:
    get:
      parameters:
      - description: Host ID
        in: query
        name: hostId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取用户可用的系统用户列表
      tags:
      - system-users
  /api/system-users/check-permission:
    get:
      parameters:
      - description: Host ID
        in: query
        name: hostId
        required: true
        type: string
      - description: System User ID
        in: query
        name: systemUserId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 检查用户权限
      tags:
      - system-users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"